Starting out with port of older Clojure utility
This can render a grid of glyphs with approximately-uniform width into
exactly-uniform-size cells. It is in black and white with no
anti-aliasing so the distance field font effect can be applied to it,
but there's no technical reason AA can't be applied if distance fields
aren't in use. They should be soon. This will also need to generate a
.fnt file, which might be the trickiest part.