Smaller font textures in premade/limited_size
The new ones are almost the same as the old ones, but are 2048x2048
(with 3 pngs) instead of 4096x4096 (with one png). The font size is
probably not exactly the same, and may be slightly smaller or slightly
larger (depending on how much the reduction in padding affects it).
There may be less glyphs, I'm not sure, but Chinese and Japanese are
supported as far as the Unicode Basic Multilingual Plane allows, plus
this has some Cyrillic and Greek in addition to good Latin script
support.