Generate and store multi-channel SDF fonts
These are in the premade/msdf folder, and if you can run Glamer
(Windows-only for now because I don't have a way to build msdfgen for
Linux and Mac), then you could generate more. The site for msdfgen is
https://github.com/Chlumsky/msdfgen . I have no idea how well these
files work because I don't have a shader to test them with yet.