Improve MSDF generation, update msdfgen
Now Glamer includes all 3 architectures of msdfgen binary, as used by
msdf-bmfont-xml , and should call the appropriate one on non-Windows
platforms. The handling for MSDF fonts should be much better, with all
the Iosevka MSDF versions lining up correctly in SquidLib testing. This
needed a lot of trial and error.